restructuring, add recipe card from other project
