+++ src/app/recipe/recipe-card/recipe-card.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-recipe-card',
+  standalone: true,
+  imports: [],
+  templateUrl: './recipe-card.component.html',
+  styleUrl: './recipe-card.component.css'
+})
+export class RecipeCardComponent {
+
+}
