+++ src/app/ui/footer/footer.component.spec.ts
