+++ src/app/ui/footer/footer.component.ts
